+++ cpp/ql/src/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 1.9.0
+
+### New Queries
+
+* Added a new query, `cpp/ambiguous-assignment-of-comparison`, to detect potentially
+  ambiguous expressions where a comparison result is assigned to a variable and the
+  assignment is used as a truth value.
+
 ## 1.8.3
 
 ### Minor Analysis Improvements