+++ go/ql/src/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.6.11
+
+No user-facing changes.
+
 ## 1.6.10
 
 No user-facing changes.