+++ rust/ql/src/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.1.43
+
+No user-facing changes.
+
 ## 0.1.42
 
 ### New Queries